rename re-exporting files for test suites
