+++ tests/test-suite/test-app/callbacks/encrypt-password.ts
@@ -1,3 +1,3 @@
-import { Operation } from "../jsonapi-ts";
+import { Operation } from "../kurier";
 
 export default async (op: Operation) => ({ password: op.data.attributes.password });
