+++ tests/test-suite/test-app/callbacks/login.ts
@@ -1,4 +1,4 @@
-import { Operation, ResourceAttributes } from "../jsonapi-ts";
+import { Operation, ResourceAttributes } from "../kurier";
 
 export default async (op: Operation, user: ResourceAttributes) => {
   return op.data.attributes.email === user.email && op.data.attributes.password === user.password;
