+++ tests/test-suite/test-app/kurier.ts
@@ -1,2 +1,1 @@
 export * from "../../../src";
-
