+++ tests/test-suite/test-app/processors/article.ts
@@ -1,4 +1,4 @@
-import { KnexProcessor, HasId } from "../jsonapi-ts";
+import { KnexProcessor, HasId } from "../kurier";
 import Article from "../resources/article";
 import Vote from "../resources/vote";
