+++ tests/test-suite/test-app/resources/vote.ts
@@ -1,4 +1,4 @@
-import { Resource } from "../jsonapi-ts";
+import { Resource } from "../kurier";
 import User from "./user";
 import Article from "./article";
